requirements.txt
bidict==0.21.3
blinker==1.4
cairocffi==1.3.0
CairoSVG==2.5.2
cffi==1.14.6
click==8.0.3
cssselect2==0.4.1
defusedxml==0.7.1
Flask==2.0.2
Flask-Mail==0.9.1
Flask-SocketIO==5.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
Pillow==8.3.2
pycparser==2.20
python-dotenv==0.19.1
python-engineio==4.2.1
python-socketio==5.4.0
tinycss2==1.1.0
webencodings==0.5.1
Werkzeug==2.0.2
